move shared controller code to embedded crate
