+++ spooky-embedded/src/controllers/embedded.rs
